api/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
python-dotenv
PyPDF2
scikit-learn
openrouter
